+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -10,6 +10,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 - Added the ability to register custom WordPress REST API controllers/endpoints via the service container
 - Added the ability to execute queries as a DRY RUN, marked with a `[DRY-RUN]` prefix in the logs
+- Added the "External Tools" settings category, for extensions to surface tooling/informational sections
 
 ### Improvements